Merging reference-relocation-branch. READ WARNING AT END
This is important if:
	You are using 1.3 with References currently
	You are upgrading from 1.2 with References

References in 1.2 were store as simple uid mappings in the Archetypes Tool
In 1.3 I introduced more complex reference objects and keep those in a folderish catalog.
While a great improvement in terms of flexibility both the 1.2 and the 1.3 model had problems
with partial site import/exports.

To address this to some extent we now store the 1.3 styled references object on the source Object of the reference.
Thus when a source object is exported it carries with it the UUID of the objects it will seek to reference. It is however
still indexed in a catalog of its own for efficency. Additionally the paths in the UID and Reference catalog have been moved
from absolute style paths to portal rooted relative paths, again to aid in import/export.


WARNING:

A migration exist in Extensions/migrations from which you can create an external method calling migrate.
This should detect 1.2 or older 1.3 style references and move them to the correct place. However you will want to create
a backup of your ZODB before attempting this and then verify that its working.