AT test suite cleanup, part I.
Please dont call self.getPortal() from tests. While it seemingly works the
idea is to use the self.portal attribute instead. The test user's home folder
can be accessed via self.folder. If content objects are created in self.folder
there is no need to switch to Manager privileges.

ArcheSiteTestCase still provides a manager user. To login as manager user call
self.login('manager'). The getManagerUser() method is gone as are the calls to
newSecurityManger(). PloneTestCase handles all of this for us.